Renamed SecondWalkerAttempt to Walker for consistancy
